+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [#961] New hook `useAsyncOptionList` which augments a `ValueSelectorProps` or `ValueEditorProps` object with async option loading and (optionally) caching.
 
+### Fixed
+
+- [#974] `parseCEL` correctly handles custom expressions in subqueries.
+
 ## [v8.11.2] - 2025-11-06
 
 ### Fixed
@@ -2063,6 +2067,7 @@ _(This list may look long, but the breaking changes should only affect a small m
 [#961]: https://github.com/react-querybuilder/react-querybuilder/pull/961
 [#967]: https://github.com/react-querybuilder/react-querybuilder/pull/967
 [#969]: https://github.com/react-querybuilder/react-querybuilder/pull/969
+[#974]: https://github.com/react-querybuilder/react-querybuilder/pull/974
 
 <!-- #endregion -->