+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+N/A
+
+## [v8.12.0] - 2025-11-06
+
 ### Added
 
-- [#961] New hook `useAsyncOptionList` which augments a `ValueSelectorProps` or `ValueEditorProps` object with async option loading and (optionally) caching.
+- [#961] New hook `useAsyncOptionList` which augments a `ValueSelectorProps` or `ValueEditorProps` object with async option list loading and caching.
 
 ### Fixed
 
@@ -2073,7 +2077,8 @@ _(This list may look long, but the breaking changes should only affect a small m
 
 <!-- #region Release comparison links -->
 
-[unreleased]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.11.2...HEAD
+[unreleased]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.12.0...HEAD
+[v8.12.0]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.11.2...v8.12.0
 [v8.11.2]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.11.1...v8.11.2
 [v8.11.1]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.11.0...v8.11.1
 [v8.11.0]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.10.0...v8.11.0